requirements.txt
backports-abc==0.5
backports.weakref==1.0rc1
bleach==1.5.0
certifi==2017.4.17
chardet==3.0.4
CurrencyConverter==0.13
cycler==0.10.0
decorator==4.1.1
entrypoints==0.2.3
flake8==3.3.0
html5lib==0.9999999
idna==2.5
ipykernel==4.6.1
ipython==6.1.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
jedi==0.10.2
Jinja2==2.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.1.0
jupyter-core==4.3.0
mahotas==1.4.3
Markdown==2.6.8
MarkupSafe==1.0
matplotlib==2.0.2
mccabe==0.6.1
mistune==0.7.4
nbconvert==5.2.1
nbformat==4.3.0
notebook==5.0.0
numpy==1.13.1
olefile==0.44
pandas==0.20.3
pandocfilters==1.4.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.2.1
prompt-toolkit==1.0.14
protobuf==3.3.0
ptyprocess==0.5.2
pycodestyle==2.3.1
pyflakes==1.5.0
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
pyzmq==16.0.2
qtconsole==4.3.0
requests==2.18.1
scikit-learn==0.18.2
scipy==0.19.1
simplegeneric==0.8.1
six==1.10.0
tensorflow-gpu==1.2.1
terminado==0.6
testpath==0.3.1
tornado==4.5.1
tqdm==4.14.0
traitlets==4.3.2
typing==3.6.1
urllib3==1.21.1
wcwidth==0.1.7
Werkzeug==0.12.2
widgetsnbextension==2.0.0
